requirements.txt
# HTTP and Async
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Dynamic Content Scraping
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Database
pymongo>=4.6.0
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Retry and Rate Limiting
tenacity>=8.2.0
ratelimit>=2.2.1

# User Agent and Proxies
fake-useragent>=1.4.0
requests[socks]>=2.31.0

# Progress and CLI
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0

# Utilities
python-dateutil>=2.8.2
validators>=0.22.0
urllib3>=2.1.0

# Scheduling (optional)
schedule>=1.2.0
APScheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.24.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Type Stubs
types-requests>=2.31.0
types-PyYAML>=6.0.0
